refactor: Extract display logic from CLI into dedicated module
Separates presentation logic from CLI command handling by moving all
Rich console formatting functions to a new display.py module.

Changes:
- Created taskcards_monitor/display.py with all display functions
- Moved create_table(), display_changes(), display_state() to display.py
- Added display_boards_list() for the list command output
- Added display_inspect_header() and display_inspect_results() for inspect command
- Updated cli.py to import and use display functions
- Removed unused Console import from cli.py

Benefits:
- Better separation of concerns (CLI logic vs presentation)
- Display logic can be reused or tested independently
- cli.py is now 192 lines vs 410 lines (53% reduction)
- Easier to modify display formatting without touching CLI logic
